Fixed some errors with MakeCfgClassesNull
Also Added math function

+++ TinManXI/src/org/usfirst/frc/team1922/robot/subsystems/BallRetrieverMotor.java
@@ -29,7 +29,8 @@ public class BallRetrieverMotor extends Subsystem {
     
     public void MakeCfgClassesNull()
     {
-    	mRetrievalMotor.delete();
+    	if(mRetrievalMotor != null)
+    		mRetrievalMotor.delete();
     	mRetrievalMotor = null;
     }
 }
